requirements.txt
langchain>=0.0.208
pdfminer.six>=20201018
pytesseract>=0.3.10
tabula-py>=2.5.1
transformers>=4.33.0
torch>=1.13.1
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
pandas>=1.3.5
numpy>=1.21.6
pdf2image>=1.16.3
openai>=0.28.0
langchain_community
flask
flask_cors
tiktok
python-docx
